\name{shinyMethyl-package}
\alias{shinyMethyl-package}
\alias{shinyMethyl}
\docType{package}
\title{
Interactive visualization for Illumina methylation arrays
}
\description{
Interactive tool for visualizing Illumina methylation array data (450k and EPIC arrays). 
}
\author{
Jean-Philippe Fortin <jfortin@jhsph.edu>
}
\keyword{ package }
