\name{shinyMethyl-package}
\alias{shinyMethyl-package}
\alias{shinyMethyl}
\docType{package}
\title{
Interactive visualization for Illumina's 450k methylation arrays
}
\description{
Interactive tool for visualizing Illumina's 450k array data
}
\author{
Jean-Philippe Fortin <jfortin@jhsph.edu>
}
\keyword{ package }

